backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
gunicorn
django-cors-headers
celery
redis
boto3
python-dotenv
dj-database-url
requests
